Added option to disable SQL Backend for commons (rendering it strictly as a library) Began to put safe-holds in place so that if Commons is being used as only a library, none of the SQL methods / operations would cause an issue.

+++ src/main/java/com/caved_in/commons/listeners/PlayerQuitListener.java
@@ -13,7 +13,11 @@ public class PlayerQuitListener implements Listener {
 		if (!Commons.getConfiguration().getWorldConfig().isJoinLeaveMessagesEnabled()) {
 			event.setQuitMessage(null);
 		}
+
 		Players.removeData(playerName);
-		Commons.disguiseDatabase.deletePlayerDisguiseData(playerName);
+
+		if (Commons.hasSqlBackend()) {
+			Commons.disguiseDatabase.deletePlayerDisguiseData(playerName);
+		}
 	}
 }
